MatrixPilot_FreeRTOS: initial clean up and update of old branch in order to reuse for latest RTOS usage exploration.
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_FreeRTOS@3109 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/airspeedCntrl.c
@@ -25,44 +25,44 @@
 
 #include "airspeedCntrl.h"
 
-extern int desiredSpeed;
+extern int16_t desiredSpeed;
 
-int 	airspeed		= 0;
-int 	groundspeed		= 0;
-int 	airspeedError	= 0;
+int16_t 	airspeed		= 0;
+int16_t 	groundspeed		= 0;
+int16_t 	airspeedError	= 0;
 
-int minimum_groundspeed		= MINIMUM_GROUNDSPEED * 100.0;
-int minimum_airspeed		= MINIMUM_AIRSPEED * 100.0;
-int maximum_airspeed		= MAXIMUM_AIRSPEED * 100.0;
+int16_t minimum_groundspeed		= MINIMUM_GROUNDSPEED * 100.0;
+int16_t minimum_airspeed		= MINIMUM_AIRSPEED * 100.0;
+int16_t maximum_airspeed		= MAXIMUM_AIRSPEED * 100.0;
 
 // Calculate the airspeed.
 // Note that this airspeed is a magnitude regardless of direction.
 // It is not a calculation of forward airspeed.
 void calc_airspeed(void)
 {
-	int speed_component ;
-	long fwdaspd2;
+	int16_t speed_component;
+	int32_t fwdaspd2;
 
-	speed_component = IMUvelocityx._.W1 - estimatedWind[0] ;
-	fwdaspd2 = __builtin_mulss ( speed_component , speed_component ) ;
+	speed_component = IMUvelocityx._.W1 - estimatedWind[0];
+	fwdaspd2 = __builtin_mulss (speed_component, speed_component);
 
-	speed_component = IMUvelocityy._.W1 - estimatedWind[1] ;
-	fwdaspd2 += __builtin_mulss ( speed_component , speed_component ) ;
+	speed_component = IMUvelocityy._.W1 - estimatedWind[1];
+	fwdaspd2 += __builtin_mulss (speed_component, speed_component);
 
-	speed_component = IMUvelocityz._.W1 - estimatedWind[2] ;
-	fwdaspd2 += __builtin_mulss ( speed_component , speed_component ) ;
+	speed_component = IMUvelocityz._.W1 - estimatedWind[2];
+	fwdaspd2 += __builtin_mulss (speed_component, speed_component);
 
 	airspeed  = sqrt_long(fwdaspd2);
 }
 
 // Calculate the groundspeed
-void calc_groundspeed(void) // computes (1/2gravity)*( actual_speed^2 - desired_speed^2 )
+void calc_groundspeed(void) // computes (1/2gravity)*(actual_speed^2 - desired_speed^2)
 {
-	long gndspd2;
+	int32_t gndspd2;
 
-	gndspd2 = __builtin_mulss ( IMUvelocityx._.W1 , IMUvelocityx._.W1 ) ;
-	gndspd2 += __builtin_mulss ( IMUvelocityy._.W1 , IMUvelocityy._.W1 ) ;
-	gndspd2 += __builtin_mulss ( IMUvelocityz._.W1 , IMUvelocityz._.W1 ) ;
+	gndspd2 = __builtin_mulss (IMUvelocityx._.W1, IMUvelocityx._.W1);
+	gndspd2 += __builtin_mulss (IMUvelocityy._.W1, IMUvelocityy._.W1);
+	gndspd2 += __builtin_mulss (IMUvelocityz._.W1, IMUvelocityz._.W1);
 
 	groundspeed 	= sqrt_long(gndspd2);
 }
@@ -70,9 +70,9 @@ void calc_groundspeed(void) // computes (1/2gravity)*( actual_speed^2 - desired_
 // Calculate the required airspeed in cm/s.  desiredSpeed is in dm/s
 void calc_target_airspeed(void)
 {
-	union longww accum ;
-	accum.WW = __builtin_mulsu ( desiredSpeed , 10 ) ;
-	target_airspeed = accum._.W0 ;
+	union longww accum;
+	accum.WW = __builtin_mulsu (desiredSpeed, 10);
+	target_airspeed = accum._.W0;
 
 	if(groundspeed < minimum_groundspeed)
 		target_airspeed += (minimum_groundspeed - groundspeed);
@@ -85,7 +85,7 @@ void calc_target_airspeed(void)
 
 	//Some airspeed error filtering
 	airspeedError = airspeedError >>=1;
-	airspeedError += ( (target_airspeed - airspeed) >> 1);
+	airspeedError += ((target_airspeed - airspeed) >> 1);
 }
 
 #endif		//(ALTITUDE_GAINS_VARIABLE == 1)
